### 2022-03-05 #4224 @editorialbot
[PRE REVIEW]: rbmi: A R package for standard and reference-based multiple imputation methods
**Submitting author:** @nociale (<a href="http://orcid.org/0000-0002-0110-7358">Alessandro Noci</a>)
**Repository:** https://github.com/insightsengineering/rbmi
**Branch with paper.md** (empty if default branch):
**Version:** 1.1.0
