### 2022-03-05 #4224 @Kevin-Mattheus-Moerman
@nociale thanks for this submission. I am the AEiC on call for the coming week and will be looking for a handling editor. In the meantime perhaps you can address those potentially missing DOIs :point_up: , you can update the paper by calling `@editorialbot generate pdf` and you can use `@editorialbot check references` to check those DOIs again.
