+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: rbmi
 Title: Reference Based Multiple Imputation
-Version: 1.1.1
+Version: 1.1.2
 Authors@R: c(
     person("Craig", "Gower-Page", email = "craig.gower-page@roche.com", role = c("aut", "cre")),
     person("Alessandro", "Noci", email = "alessandro.noci@roche.com", role = c("aut")),
