+++ DESCRIPTION
@@ -11,6 +11,8 @@ Description: Implements reference based multiple imputation allowing for the imp
 Encoding: UTF-8
 LazyData: true
 Roxygen: list(markdown = TRUE)
+URL: https://insightsengineering.github.io/rbmi/, https://github.com/insightsengineering/rbmi
+BugReports: https://github.com/insightsengineering/rbmi/issues
 RoxygenNote: 7.1.2
 Suggests: 
     dplyr,
