### 2022-03-16 4cf7203
added index limit to stan model
