+++ man/as_indices.Rd
@@ -15,6 +15,6 @@ Converts a string of 0's and 1's into index positions of the 1's
 padding the results by 0's so they are all the same length
 }
 \details{
-i.e.\preformatted{patmap(c("1101", "0001"))  ->   list(c(1,2,4,0), c(4,0,0,0))
+i.e.\preformatted{patmap(c("1101", "0001"))  ->   list(c(1,2,4,999), c(4,999, 999, 999))
 }
 }
