### 2022-03-17 #4224 @fboehm
Thanks, @DanielRivasMD ! If you decide that you'd like to review this, please let me know. I'm happy to guide you through the process of reviewing for JOSS. You would need to install the R package and verify its functionality. You would also need to read the accompanying paper and ensure that it meets specified expectations. If you decide to review this submission, I'll prepare a standard checklist for you to complete. Unlike many journals, you are not expected to write a document that contains your review for the JOSS submission. Instead, you merely complete a checklist. A JOSS review is an iterative process, so you'll first download and install the R package and try to run it. Perhaps something won't work as indicated, in which case you'd tell the authors about it via our review thread. Nearly every submission that gets reviewed eventually gets published, once the reviewers are satisfied with any needed modifications. I hope that this helps.
