### 2022-03-17 #4224 @fboehm
To get a sense of the checklist items, please see this ongoing review: https://github.com/openjournals/joss-reviews/issues/4242
