### 2022-03-17 #4224 @DanielRivasMD
Sounds great! Let's do it! I read the article, and will install and test the software, is this the review thread where I should comment? Or where can I find it?
