+++ misc/joss_manuscript/references.bib
@@ -249,7 +249,7 @@
 
 @article{Mallinckrodt2008,
   title={Recommendations for the primary analysis of continuous endpoints in longitudinal clinical trials},
-  author={Mallinckrod, Craig H and Lane, Peter W and Schnell, Dan and Peng, Yahong and Mancuso, James P},
+  author={Mallinckrodt, Craig H and Lane, Peter W and Schnell, Dan and Peng, Yahong and Mancuso, James P},
   journal={Drug Information Journal},
   volume={42},
   number={4},
@@ -425,7 +425,7 @@ protocol deviation},
 
 @misc{Wolbers2021,
   author = {Wolbers, Marcel and Noci, Alessandro and Delmar, Paul and Gower-Page, Craig and Yiu, Sean and Bartlett, Jonathan W},
-  title = {Reference-based imputation methods based on conditional mean imputation},
-  year = {2021},
+  title = {Standard and reference-based conditional mean imputation},
+  year = {2022},
   url={https://arxiv.org/abs/2109.11162}
 }
