+++ pkgdown.yml
@@ -5,5 +5,5 @@ articles:
   advanced: advanced.html
   quickstart: quickstart.html
   stat_specs: stat_specs.html
-last_built: 2022-04-08T12:22Z
+last_built: 2022-04-08T12:42Z
 
