### 2022-04-08 91b5092
renamed R_TEST_NIGHTLY to R_TEST_FULL
