+++ .github/workflows/on_pr_cran.yaml
@@ -26,7 +26,7 @@ jobs:
 
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
-      R_TEST_NIGHTLY: 'FALSE' 
+      R_TEST_FULL: 'FALSE' 
 
     steps:
       - name: Checkout
