+++ .github/workflows/on_weekly.yaml
@@ -1,68 +1,0 @@
-
-
-on:
-  schedule:
-    - cron: '0 4 * * 1'
-
-name: On Weekly
-
-jobs:
-  
-  rcmdcheck-cran:
-    strategy:
-      fail-fast: true
-      matrix:
-       config:
-         - { image: "ghcr.io/${{ github.repository_owner }}/rbmi:r404"}
-         - { image: "ghcr.io/${{ github.repository_owner }}/rbmi:r410"}
-
-    runs-on: ubuntu-latest
-
-    container:
-      image: ${{ matrix.config.image }}
-
-    env:
-      GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
-      R_TEST_FULL: 'TRUE' 
-
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v2
-        with:
-          ref: cran
-
-      - name: Check
-        uses: ./.github/actions/rcmdcheck
-        with:
-          slim: true
-  
-  
-  
-  rcmdcheck-main:
-    strategy:
-      fail-fast: true
-      matrix:
-       config:
-         - { image: "ghcr.io/${{ github.repository_owner }}/rbmi:r404"}
-         - { image: "ghcr.io/${{ github.repository_owner }}/rbmi:r410"}
-
-    runs-on: ubuntu-latest
-
-    container:
-      image: ${{ matrix.config.image }}
-
-    env:
-      GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
-      R_TEST_FULL: 'TRUE' 
-
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v2
-        with:
-          ref: main
-
-      - name: Check
-        uses: ./.github/actions/rcmdcheck
-        with:
-          slim: true
-
