+++ .github/workflows/build_docker_latest.yaml
@@ -5,7 +5,7 @@ env:
 
 on:
   schedule:
-    - cron: '0 5 1,15 * *'
+    - cron: '0 5 1 * *'
   workflow_dispatch:
 
 
