+++ .github/workflows/on_biweekly.yaml
@@ -3,6 +3,7 @@
 on:
   schedule:
     - cron: '0 4 1,15 * *'
+  workflow_dispatch:
 
 name: Bi-Weekly
 
