### 2022-04-10 #4251 @DanielRivasMD
@fboehm Hi, my apologies it took so long to get back to this. Now, I have taken a deeper look into the source code (I had already read the paper and looked at the repo superficially) and tested the functionality. Overall, I think it is a pretty complete package, extensibly tested, good documentation and a use case attached as a dataset (indicated in the checklist that I hope you can see above ^). I have only three issues with this project, and not entirely certain whether they fit with JOSS criteria:

1) Community / contributors guidelines. There is only a link at the README to open issues, but no suggestions on format or further documentation.
2) The paper does not present state of the field for comparison with other software.
3) In order to use and test, dependencies must be installed. I would have like this to be documented at least if not automatically setup.

Please do not hesitate to come back if any point is unclear, or further issues must be discussed.
