### 2022-04-10 #4251 @fboehm
Thank you, @DanielRivasMD ! I think that you make some good points in the comment. @nociale, please address the points that @DanielRivasMD made above:

1. Community / contributors guidelines. There is only a link at the README to open issues, but no suggestions on format or further documentation.
1. The paper does not present state of the field for comparison with other software.
1. In order to use and test, dependencies must be installed. I would have like this to be documented at least if not automatically setup.

Thank you!
