### 2022-04-11 #4251 @JoranTiU
@fboehm Hi, I agree with @DanielRivasMD, it is a nice and pretty complete package with good documentation. The Vigenette's are also very helpful. Next to the points raised by @DanielRivasMD (i.e., (i) There is only a link at the README to open issues, but no suggestions on format or further documentation; (ii) The paper does not present state of the field for comparison with other software, and (iii) In order to use and test, dependencies must be installed.) I had  the following minor remarks:

1. The examples provided with the functions can’t be directly run by the user (e.g.,  in the description of the “impute” function  the example can’t be executed because the drawobj used in the example does not exist). This makes the examples less useful and also makes it harder to verify the package is working properly?
2.  In description of “draws” function: “The imputation model is a mixed effects model repeated measures (MMRM) model.” Remove first instance of the word model.
3.  In description of “draws” function: “It can be fit using frequentist maximum likelihood (ML) or restricted ML (REML)”. REML is also frequentist.
4. In quickstart vignette: “The analysis model is an ANCOVA model with the treatment group as the primary covariate and adjustment for the baseline HAMD17 score.” This suggests group is the covariate while it is the factor.

Overall though, as I said, like the package :).

Best,
Joran
