### 2022-04-11 92fe5b7
Update CONTRIBUTING.md
Spelling and re-wording
