+++ CONTRIBUTING.md
@@ -1,6 +1,6 @@
-# Contributing to rbmi
+# Contributing to `rbmi`
 
-This file outlines how to propose and make changes to rbmi as well as providing details about more obscure aspects of the packages development process.
+This file outlines how to propose and make changes to `rbmi` as well as providing details about more obscure aspects of the package’s development process.
 
 
 ## Code changes
@@ -11,11 +11,11 @@ If you’ve found a bug, please file an issue that illustrates the bug with a mi
 
 ### Pull request process
 
-* This project uses a simple GitHub flow model for development. That is, code changes should be done in their own feature branch based off of "main" and merged back into "main" once complete.
+* This project uses a simple GitHub flow model for development. That is, code changes should be done in their own feature branch based off of the `main` branch and merged back into the `main` branch once complete.
 
 * Pull Requests will not be accepted unless all CI/CD checks have passed. (See the CI/CD section for more information).
 
-* Pull Requests relating to any of the packages core R code must be accompanied by a corresponding unit test. Any new features or bug fixes will be rejected if they do not also contain a unit test to demonstrate that they are working as intended. (See the Unit Testing section for more information).
+* Pull Requests relating to any of the package's core R code must be accompanied by a corresponding unit test. Any pull requests containing changes to the core R code that do not contain a unit test to demonstrate that it is working as intended will not be accepted. (See the Unit Testing section for more information).
 
 * Pull Requests should add a few lines about what has been changed to the `NEWS.md` file.
 
@@ -24,13 +24,13 @@ If you’ve found a bug, please file an issue that illustrates the bug with a mi
 
 *  We use [roxygen2](https://cran.r-project.org/package=roxygen2), with [Markdown syntax](https://cran.r-project.org/web/packages/roxygen2/vignettes/rd-formatting.html), for documentation.  
 
-* Please ensure your code conforms to `lintr`. You can check this by running `lintr::lint("FILE NAME")` on any files you have modified and ensuring that the findings are kept to as few as possible. (We do not have any hard requirements on following lintr but do encourage developers to follow its guidance as closely as possible)
+* Please ensure your code conforms to [lintr](https://cran.r-project.org/web/packages/lintr/readme/README.html#available-linters). You can check this by running `lintr::lint("FILE NAME")` on any files you have modified and ensuring that the findings are kept to as few as possible. We do not have any hard requirements on following `lintr`'s conventions but do encourage developers to follow its guidance as closely as possible.
 
-* This project uses 4 space indents, contributions not following this will be rejected. 
+* This project uses 4 space indents, contributions not following this will not be accepted. 
 
-* This project makes use S3 of R6 for OOP. Usage of S4 and other OOP systems is to be avoided unless absolutely necessary.
+* This project makes use of [S3](https://adv-r.hadley.nz/s3.html) and [R6](https://adv-r.hadley.nz/r6.html) for OOP. Usage of S4 and other OOP systems should be avoided unless absolutely necessary to ensure consistency. Having said that it is recommended to stick to S3 unless modification in place or other R6 specific features are required.
 
-* The current desire of this package is to keep the dependency tree as small as possible. To that end you are discouraged from adding any additional packages to the "Depends" / "Imports" section unless absolutely essential. If you are importing a package just to use a single function consider writing a variant of that function with base R. There are no such expectations for "Suggests" and you are free to use any package for the vignettes / unit tests (though again please be mindful to not be unnecessarily excessive with this).
[…]
