+++ NEWS.md
@@ -1,0 +1,14 @@
+# rbmi (development version)
+
+* Updated reference urls
+
+  
+# rbmi 1.1.1 & 1.1.3
+  
+* Various minor tweaks to address CRAN checks
+* Otherwise no change in functionality from 1.1.0
+  
+# rbmi 1.1.0
+  
+* Initial public release
+  
