### 2022-04-24 #4251 @DanielRivasMD
Hi all (@fboehm @nociale), apologies for the delay in this reply. In reply to the comments above:

1) I believe that adding such files to the community helps enormously to make this package a collective effort. I applaud such a resolution.
2) As a matter of fact, initially I did not review the references that the authors indicate above. However, they, in my personal opinion, provide enough evidence for comparison between this and other implementations. I leave it to the editors to decide whether this particular aspect is in coherence with the journal policy.
3) When I installed `rbmi` via `install.packages()` it errored due to a dependency, which is not unheard of in `R`. Nevertheless, I think that the newly provided documentation should suffice.

Thanks.
