### 2022-04-25 #4251 @JoranTiU
Hi everyone (@fboehm @nociale) !

Apologies for my late reply too, I was on holiday last week ;).

I'm very happy and very satisfied with the changes :). The runnable examples would be nice, but the Vignette makes up for this to a large extend, so not having those definitely isn't a deal breaker as far as I'm concerned :).

Thanks for all the work and the great package :).

Best,
Joran
