### 2022-04-27 782d08b
add change function and lst2df function to utility
