### 2022-04-27 670d82e
wrap pool to return data.frame instead of list
