+++ pkgdown.yml
@@ -1,9 +1,9 @@
 pandoc: 2.17.1.1
-pkgdown: 2.0.2
+pkgdown: 2.0.3
 pkgdown_sha: ~
 articles:
   advanced: advanced.html
   quickstart: quickstart.html
   stat_specs: stat_specs.html
-last_built: 2022-04-20T11:14Z
+last_built: 2022-05-04T11:11Z
 
