### 2022-05-04 dacf6e9
reworked get_data() code
