### 2022-05-06 4150c76
use analysis_results object in ancova instead of named list
