### 2022-05-06 6bf4538
update analysis function to handle analysis_result object
