+++ R/analyse.R
@@ -341,6 +341,16 @@ print.analysis <- function(x, ...) {
         n_samp
     )
 
+    info <- function(where, what) {
+        if (all(length(where) != 0,
+                !is.null(where[[1]]),
+                rlang::has_name(where[[1]], what))) {
+            sapply(where, `[[`, what)
+        } else {
+            paste("No", what, "info")
+        }
+    }
+
     string <- c(
         "",
         "Analysis Object",
@@ -349,7 +359,7 @@ print.analysis <- function(x, ...) {
         sprintf("Analysis Function: %s", x$fun_name),
         sprintf("Delta Applied: %s", !is.null(x$delta)),
         "Analysis Estimates:",
-        sprintf("    %s", names(x$results[[1]])),
+        sprintf("    %s", info(x$results[[1]], 'name')),
         ""
     )
 
@@ -440,9 +450,9 @@ validate_analyse_pars <- function(results, pars) {
     )
 
     assert_that(
-        length(names(results[[1]])) != 0,
-        all(vapply(results, function(x) !is.null(names(x)) & all(names(x) != ""), logical(1))),
-        msg = "Individual analysis results must be named lists"
+        length(results[[1]]) != 0,
+        all(vapply(results, function(Xs) all(vapply(Xs, function(X) is.analysis_result(X), logical(1))), logical(1))),
+        msg = "Individual analysis results must be type of analysis_result"
     )
 
     results_names <- lapply(results, function(x) unique(names(x)))
@@ -486,4 +496,181 @@ validate_analyse_pars <- function(results, pars) {
     }
 
     return(invisible(TRUE))
+}
+
+#' Constructor of analysis result
+#'
+#' Construct an analysis result class object whose base type is a list
+#'
+#' @param name A character variable for the group name
+#' @param est A double type numeric variable as the estimate
+#' @param se A double type numeric variable as the standard error
+#' @param df An integer type of numeric variable
+#' @param meta A list type of variable as meta information
+#' @details
+#' - `se` must be numeric values greater or equal to 0
+#' - `meta` is optional
+#' @return An object of "analysis_result" class
+#' @examples
+#' \dontrun{
+#' ana_res_obj <- analysis_result(name = 'trt', est = 1, se = 2, df = as.integer(3), meta = list(visit = 1))
+#' }
+#' @export
+analysis_result <- function (name = character(),
+                             est = double(),
+                             se = double(),
+                             df = integer(),
+                             meta = NULL) {
+
+    # constraints
+    stopifnot(is.character(name))
+    stopifnot(is.double(est))
+    stopifnot(is.double(se))
+    stopifnot(is.integer(df) | is.double(df))
+    stopifnot(is.list(meta) | is.null(meta))
+
+    # validators
+    if (se < 0) stop("SE must great or equal to 0", .call = FALSE)
+
+    value <- list(name = name,
+                  est = est,
+                  se = se,
+                  df = df)
+
+    if (!is.null(meta)) {
+        value[['meta']] <- meta
+    }
+
+    structure(
+        value,
+        meta = meta,
+        class = "analysis_result"
+    )
+}
+
+#' Convert object to analysis result class
+#'
+#' @param x The object to be converted to analysis_result class
+#' @param ... Optional keywords parameters for adding missing elements to the object
+#' @return An "analysis_result" class object with optionally updated elements
+#' @examples
+#' \dontrun{
+#' ana_res_obj <- as_analysis_result(list(est = 1, se = 2, df = 3), name = 'trt')
+#' }
+#' @export
+as_analysis_result <- function(x, ...) {
+    dots <- rlang::enquos(...)
+
+    # coercion with generic function
+    x <- as.list(x)
+
+    present <- ana_name_chker('present')
+
+    names_not_presented <- names(present(x))[!present(x)]
+
[…]
