### 2022-05-06 22fd021
add util function of add meta
