### 2022-05-06 caf8a2d
update example of analysis function in the header
