### 2022-05-06 6a97460
update print.analysis function to print analysis info in a tabular format
