+++ man/add_meta.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/utilities.R
+\name{add_meta}
+\alias{add_meta}
+\title{Add meta information to customerize analysis function}
+\usage{
+add_meta(var_names, var_values)
+}
+\arguments{
+\item{name}{The name of the element to be added to meta}
+
+\item{...}{The values of the element to be added to meta
+This function used only internally for ancova}
+}
+\description{
+Add meta information to customerize analysis function
+}
