+++ man/ana_name_chker.Rd
@@ -1,0 +1,14 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/analyse.R
+\name{ana_name_chker}
+\alias{ana_name_chker}
+\title{Name checker for analysis function}
+\usage{
+ana_name_chker(msg)
+}
+\arguments{
+\item{msg}{Character vector representing which checker to return}
+}
+\description{
+Name checker for analysis function
+}
