+++ man/analysis_info.Rd
@@ -1,0 +1,24 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/analyse.R
+\name{analysis_info}
+\alias{analysis_info}
+\title{Get printable analysis information from an example of analysis result}
+\usage{
+analysis_info(example, name_of_meta = "meta")
+}
+\arguments{
+\item{example}{A subset of the result of the analysis object for getting enough info to print}
+
+\item{name_of_meta}{A character variable for the name of meta data in the result of analysis. Default: 'meta'}
+}
+\value{
+A data.frame containing the information of the analysis result from the example
+}
+\description{
+The example should not be the complete result of analysis object but a subset of it such as \code{anaObj$results[[1]]}
+}
+\examples{
+\dontrun{
+analysis_info(dat, name_of_meta = 'meta')
+}
+}
