+++ man/namechecker.Rd
@@ -1,0 +1,19 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/analyse.R
+\name{namechecker}
+\alias{namechecker}
+\title{Create name checkers with message passing dispatch}
+\usage{
+namechecker(..., optional = NULL)
+}
+\arguments{
+\item{...}{Character vectors for the reference to check against}
+
+\item{optional}{Character vector of optional name. Default: NULL}
+}
+\value{
+A constructor to create checker functions with message passing dispatch
+}
+\description{
+Create name checkers with message passing dispatch
+}
