+++ man/analysis_result.Rd
@@ -1,0 +1,42 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/analyse.R
+\name{analysis_result}
+\alias{analysis_result}
+\title{Constructor of analysis result}
+\usage{
+analysis_result(
+  name = character(),
+  est = double(),
+  se = double(),
+  df = integer(),
+  meta = NULL
+)
+}
+\arguments{
+\item{name}{A character variable for the group name}
+
+\item{est}{A double type numeric variable as the estimate}
+
+\item{se}{A double type numeric variable as the standard error}
+
+\item{df}{An integer type of numeric variable}
+
+\item{meta}{A list type of variable as meta information}
+}
+\value{
+An object of "analysis_result" class
+}
+\description{
+Construct an analysis result class object whose base type is a list
+}
+\details{
+\itemize{
+\item \code{se} must be numeric values greater or equal to 0
+\item \code{meta} is optional
+}
+}
+\examples{
+\dontrun{
+ana_res_obj <- analysis_result(name = 'trt', est = 1, se = 2, df = as.integer(3), meta = list(visit = 1))
+}
+}
