+++ NAMESPACE
@@ -37,7 +37,9 @@ S3method(validate,stan_data)
 export(Stack)
 export(add_class)
 export(analyse)
+export(analysis_result)
 export(ancova)
+export(as_analysis_result)
 export(as_class)
 export(as_vcov)
 export(delta_template)
@@ -50,6 +52,7 @@ export(getStrategies)
 export(get_example_data)
 export(has_class)
 export(impute)
+export(is.analysis_result)
 export(locf)
 export(longDataConstructor)
 export(method_approxbayes)
@@ -72,6 +75,11 @@ import(R6)
 import(Rcpp)
 import(methods)
 importFrom(assertthat,assert_that)
+importFrom(dplyr,"%>%")
+importFrom(dplyr,bind_cols)
+importFrom(dplyr,bind_rows)
+importFrom(dplyr,left_join)
+importFrom(dplyr,select)
 importFrom(glmmTMB,VarCorr)
 importFrom(glmmTMB,fixef)
 importFrom(glmmTMB,getME)
