### 2022-05-09 e87eeee
add validation for object class and exceptoin handling
