### 2022-05-09 2f5feef
fix exception handling
