### 2022-05-09 9eba1d3
use assertthat instead of stopifnot and git rid of non-base functions (dplyr...)
