### 2022-05-10 26df540
prep for cran release
