+++ NEWS.md
@@ -1,6 +1,11 @@
 # rbmi (development version)
 
+* No changes
+# rbmi 1.1.4
+
 * Updated reference urls
+* Fixed bug where visit factor levels were re-constructed incorrectly in `delta_template()`
+* Fixed bug where the wrong visit was displayed in error message for when a visit doesn't have any data
 
   
 # rbmi 1.1.1 & 1.1.3
