+++ .github/workflows/check_for_cran.yaml
@@ -29,7 +29,7 @@ jobs:
         uses: actions/checkout@v2
 
       - name: Install R
-        uses: r-lib/actions/setup-r@v1
+        uses: r-lib/actions/setup-r@v2
         id: install-r
         with:
           r-version: ${{ matrix.config.R }}
@@ -40,7 +40,7 @@ jobs:
         run: |
           echo 'options(install.packages.check.source = "no")' >>  .Rprofile
 
-      - uses: r-lib/actions/setup-r-dependencies@v1
+      - uses: r-lib/actions/setup-r-dependencies@v2
         name: Install R package dependencies
         with:
           cache-version: 2
@@ -48,7 +48,7 @@ jobs:
             remotes
 
       - name: Install Pandoc
-        uses: r-lib/actions/setup-pandoc@v1
+        uses: r-lib/actions/setup-pandoc@v2
 
       - name: Check
         uses: ./.github/actions/rcmdcheck
