### 2022-05-10 deb1ee9
add util funs
