+++ R/analyse.R
@@ -444,7 +444,7 @@ validate_analyse_pars <- function(results, pars) {
     assert_that(
         length(results[[1]]) != 0,
         all(vapply(results, function(Xs) all(vapply(Xs, function(X) is.analysis_result(X), logical(1))), logical(1))),
-        msg = "Individual analysis results must be type of analysis_result"
+        msg = "Individual analysis result must be type of analysis_result"
     )
 
     results_names <- lapply(results, function(x) unique(names(x)))
@@ -509,26 +509,33 @@ validate_analyse_pars <- function(results, pars) {
 #' }
 #' @export
 analysis_result <- function (name = character(),
-                             est = double(),
-                             se = double(),
-                             df = integer(),
+                             est = numeric(),
+                             se = numeric(),
+                             df = NULL,
                              meta = NULL) {
 
     # constraints
-    stopifnot(is.character(name))
-    stopifnot(is.double(est))
-    stopifnot(is.double(se))
-    stopifnot(is.integer(df) | is.double(df))
-    stopifnot(is.list(meta) | is.null(meta))
+    is.numeric_or_NA <- make_chain(any, is.numeric, anyNA)
+    is.numeric_or_NA_or_NULL <- make_chain(any, is.numeric_or_NA, is.null)
+    is.list_or_null <- make_chain(any, is.list, is.null)
+
+    assert_type(name, is.character)
+    assert_type(est, is.numeric)
+    assert_type(se, is.numeric)
+    assert_type(df, is.numeric_or_NA_or_NULL)
+    assert_type(meta, is.list_or_null)
 
     # validators
-    if (se < 0) stop("SE must great or equal to 0", .call = FALSE)
+    if (se < 0) stop("SE must greater or equal to 0", .call = FALSE)
 
     value <- list(name = name,
                   est = est,
-                  se = se,
-                  df = df)
+                  se = se)
 
+    # optional values
+    if (!is.null(df)) {
+        value[['df']] <- df
+    }
     if (!is.null(meta)) {
         value[['meta']] <- meta
     }
@@ -536,7 +543,7 @@ analysis_result <- function (name = character(),
     structure(
         value,
         meta = meta,
-        class = "analysis_result"
+        class = c("analysis_result", "list")
     )
 }
 
@@ -556,7 +563,7 @@ as_analysis_result <- function(x, ...) {
     # coercion with generic function
     x <- as.list(x)
 
-    present <- ana_name_chker('present')
+    present <- ana_name_chker('musthave_in_objnames')
 
     names_not_presented <- names(present(x))[!present(x)]
 
@@ -574,14 +581,15 @@ as_analysis_result <- function(x, ...) {
     # after updating check if all required elements are presented
     stopifnot(all(present(updated_x)))
 
-    # keep only required elements and put them in defined order
-    if (length(names(updated_x)) < length(ana_name_chker('all'))) {
-        ordered_x <- updated_x[ana_name_chker('musthave')]
-    } else {
-        ordered_x <- updated_x[ana_name_chker('all')]
+    # order the list by names
+    ordered_x <- order_list_by_name(updated_x, ana_name_chker('all'))
+
+    # set attributes: meta & class
+    if ('meta' %in% names(ordered_x)) {
+        attr(ordered_x, 'meta') <- ordered_x[['meta']]
     }
 
-    as_class(ordered_x, "analysis_result")
+    as_class(ordered_x, c("analysis_result", "list"))
 }
 
 #' Create name checkers with message passing dispatch
@@ -605,28 +613,23 @@ namechecker <- function(..., optional = NULL) {
             function(x, y) f(y, x)
         }
 
-        extend <- function(v1, v2) {
-            if(is.null(v2)) v1
-            else append(v1, v2)
-        }
-
         # higher-order function to create template for checkers/validators
[…]
