+++ man/order_list_by_name.Rd
@@ -1,0 +1,24 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/utilities.R
+\name{order_list_by_name}
+\alias{order_list_by_name}
+\title{Order a named list by its names according to given character vector}
+\usage{
+order_list_by_name(L, v)
+}
+\arguments{
+\item{L}{A list to be ordered}
+
+\item{v}{A character contains the names in order}
+}
+\value{
+A list with names in order
+\dontrun{
+L_ordered <- order_list_by_name(list(a=1,b='x',c=TRUE), c("c", "a", "d", "x", "b", "t"))
+# returns a list `list(c=TRUE, a=1, b='x)`
+
+}
+}
+\description{
+Order a named list by its names according to given character vector
+}
