+++ NAMESPACE
@@ -75,11 +75,7 @@ import(R6)
 import(Rcpp)
 import(methods)
 importFrom(assertthat,assert_that)
-importFrom(dplyr,"%>%")
-importFrom(dplyr,bind_cols)
-importFrom(dplyr,bind_rows)
-importFrom(dplyr,left_join)
-importFrom(dplyr,select)
+importFrom(assertthat,has_attr)
 importFrom(glmmTMB,VarCorr)
 importFrom(glmmTMB,fixef)
 importFrom(glmmTMB,getME)
