### 2022-05-11 4c04422
update base_bind_rows and use it in analysis instead of dplyr::bind_rows
