### 2022-05-11 cde28a1
move namechecker function from analyse.R to utilities.R since it is more general
