### 2022-05-11 453f13a
use assert_that instead of stopifnot
