### 2022-05-11 02f37be
lazily evaluate namechecker to avoid name error
