+++ R/ancova.R
@@ -28,16 +28,28 @@
 #' If no value for `visits` is provided then it will be set to
 #' `unique(data[[vars$visit]])`.
 #'
-#' In order to meet the formatting standards set by [analyse()] the results will be collapsed
-#' into a single list suffixed by the visit name, e.g.:
+#' Visits as part of the meta information of the `analysis_result` object from results of [analyse()] can be accessed individually and are
+#' are displayed in a column from the `print.analysis` output such like
+#' ```
+#'  =====================================
+#'    name     est     se     df   visit
+#'  -------------------------------------
+#'     trt    -0.513  0.505   197    1
+#'     trt    -2.366  0.675   197    4
+#'   lsm_ref   7.51   0.477   197    4
+#'   lsm_alt  5.144   0.477   197    4
+#'  -------------------------------------
+#'
+#' ```
+#' Then list in analysis results has structure such as following. Each individual result is in class `analysis_result`
 #'```
 #'list(
-#'    trt_visit_1 = list(est = ...),
-#'    lsm_ref_visit_1 = list(est = ...),
-#'    lsm_alt_visit_1 = list(est = ...),
-#'    trt_visit_2 = list(est = ...),
-#'    lsm_ref_visit_2 = list(est = ...),
-#'    lsm_alt_visit_2 = list(est = ...),
+#'    trt = analysis_result(name =, est = ..., meta = list(visit=1, ...)),
+#'    lsm_ref = analysis_result(name =, est = ..., meta = list(visit=1, ...)),
+#'    lsm_alt = analysis_result(name =, est = ..., meta = list(visit=1, ...)),
+#'    trt = analysis_result(name =, est = ..., meta = list(visit=2, ...)),
+#'    lsm_ref = analysis_result(name =, est = ..., meta = list(visit=2, ...)),
+#'    lsm_alt = analysis_result(name =, est = ..., meta = list(visit=2, ...)),
 #'    ...
 #')
 #'```
