### 2022-05-11 2e282c1
use base R instead of rlang. add more paramters to function
