+++ man/base_bind_rows.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/utilities.R
+\name{base_bind_rows}
+\alias{base_bind_rows}
+\title{Convert nested list to data.frame}
+\usage{
+base_bind_rows(nestlist)
+}
+\arguments{
+\item{nestlist}{A nested list to be converted to data.frame}
+}
+\value{
+A data.frame binding each sublist as row in the data.frame and with NA filled for missing values
+}
+\description{
+Convert nested list to data.frame
+}
