+++ man/namechecker.Rd
@@ -1,8 +1,8 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/analyse.R
+% Please edit documentation in R/utilities.R
 \name{namechecker}
 \alias{namechecker}
-\title{Create name checkers with message passing dispatch}
+\title{Create name checkers for object with message passing dispatch}
 \usage{
 namechecker(..., optional = NULL)
 }
@@ -15,5 +15,5 @@ namechecker(..., optional = NULL)
 A constructor to create checker functions with message passing dispatch
 }
 \description{
-Create name checkers with message passing dispatch
+Create name checkers for object with message passing dispatch
 }
