### 2022-05-11 e6f6091
use anonymous function instead of partial function to avoid using purrr
