### 2022-05-11 23d23e1
add assert functions
