+++ R/analyse.R
@@ -524,8 +524,10 @@ analysis_result <- function (name,
     assert_type(se, is.numeric)
     assert_type(df, is.numeric_or_NA_or_NULL)
     assert_type(meta, is.list_or_null)
+    assert_anares_length(name, 1)
+    assert_anares_length(est, 1)
+    assert_anares_length(se, 1)
 
-    # validators
     assert_that(
         se >= 0,
         msg = "SE must be greater or equal to 0"
