+++ tests/testthat/test-utilities.R
@@ -309,3 +309,40 @@ test_that("make_chain", {
     expect_false(is.numeric_or_na(list()))
     expect_error(make_chain(any, is.numeric, 'a')(1))
 })
+
+test_that("order_list_by_name", {
+    expect_equal(order_list_by_name(list(a=1,b='x',c=TRUE), c("c", "a", "d", "x", "b", "t"))[[3]], 'x')
+    expect_true(names(order_list_by_name(list(t=1,v='x'), c("c", "a", "d", "x", "b", "t"))) == 't')
+    expect_true(all(names(order_list_by_name(list(t=1,v='x',z=2,m=list(), q='t', w=data.frame()), c("z", "t","q", "m"))) ==  c("z", "t","q", "m")))
+    expect_length(order_list_by_name(list(u=1,v='x'), c("c", "a", "d", "x", "b", "t")), 0)
+    expect_length(order_list_by_name(list(u=1,v='x'), c("c")), 0)
+})
+
+test_that("base_bind_rows", {
+    l1 <- list(list(a=1,b=2), list(b=3, c=4))
+    expect_equal(nrow(base_bind_rows(l1)), 2)
+    expect_equal(ncol(base_bind_rows(l1)), 3)
+    expect_equal(base_bind_rows(l1)[1, 2], 2)
+    l2 <- list(list(a=1,b=2, c=3), list(a=1, b=3, c=4), list(a=9, b=10, c=12))
+    expect_equal(nrow(base_bind_rows(l2)), 3)
+    expect_equal(base_bind_rows(l2)[2,3], 4)
+    expect_true(is.na(base_bind_rows(l1)[3,1]))
+    expect_error(base_bind_rows(list(list(a=1,b=2), list(b=3, c=data.frame()))))
+    l3 <- list(c(a=1,b=2), c(a=3, c=4))
+    expect_error(base_bind_rows(l3))
+})
+
+test_that("namechecker", {
+    chker <- namechecker('a', 'b', 'c', optional = c('d', 'e', 'f'))
+    expect_type(chker, "closure")
+    expect_equal(chker('musthave'), c('a', 'b', 'c'))
+    expect_equal(chker('optional'), c('d', 'e', 'f'))
+    expect_equal(chker('all'), c('a', 'b', 'c', 'd', 'e', 'f'))
+    expect_type(chker('musthave_in_objnames'), "closure")
+    expect_type(chker('objnames_in_musthave'), "closure")
+    expect_true(all(chker('musthave_in_objnames')(list(a=1)) == list(a=TRUE, b=FALSE, c=FALSE)))
+    expect_true(all(chker('musthave_in_objnames')(list(a=1, b = 2, c = 3, d = 4)) == list(a=TRUE, b=TRUE, c=TRUE)))
+    expect_true(all(chker('objnames_in_musthave')(list(b=2, a = 1, d = 2, e = 4, x = 5)) == list(b=TRUE, a=TRUE, d=TRUE, e = TRUE, x = FALSE)))
+    expect_true(all(chker('objnames_in_musthave')(list(f = 1, x = list(a=2))) == list(f=TRUE, x=FALSE)))
+    expect_true(all(chker('objnames_in_musthave')(list(y = 1, x = list(a=2))) == list(y=FALSE, x=FALSE)))
+})
