### 2022-05-13 2f6f9b5
use better variable name and update function header
