### 2022-05-13 ecb1e6e
udpate add_meta to change how parameter is called and make it more robust
