### 2022-05-13 0e16edb
allow NA for se and use anyNA instead of is.na to be more robust
