### 2022-05-13 dd8a6c5
change se from required to optional parameter for analysis_result
