### 2022-05-14 #4251 @fboehm
@nociale - please add dois to the cited references. I suggest removing from the bib file any references that aren't actually used in paper.md. It looks like you have 9 references in the pdf. We need only those to have accurate DOIs, if they're available. Thanks again!
