### 2022-05-16 d1166a3
updated check for cran gha
