+++ .github/workflows/check_for_cran.yaml
@@ -43,9 +43,9 @@ jobs:
       - name: Install R package dependencies
         shell: Rscript {0}
         run: |
-          options("repos" = list("CRAN" = "https://packagemanager.rstudio.com/cran/latest"))
-          install.packages("devtools")
-          devtools::install_dev_deps()
+          options("repos" = list("CRAN" = "https://www.stats.bris.ac.uk/R/"))
+          install.packages("remotes")
+          remotes::install_deps(dependencies = TRUE)
 
       - name: Install Pandoc
         uses: r-lib/actions/setup-pandoc@v2
