### 2022-05-18 943ef46
added to news; updated cran-comments
