+++ NEWS.md
@@ -1,17 +1,19 @@
 # rbmi (development version)
 
 * No changes
+
 # rbmi 1.1.4
 
-* Updated reference urls
-* Fixed bug where visit factor levels were re-constructed incorrectly in `delta_template()`
-* Fixed bug where the wrong visit was displayed in error message for when a visit doesn't have any data
+* Updated urls for references in vignettes
+* Fixed a bug where visit factor levels were re-constructed incorrectly in `delta_template()`
+* Fixed a bug where the wrong visit was displayed in the error message for when a specific visit doesn't have any data in `draws()`
+* Fixed a bug where the wrong input parameter was displayed in an error message in `simulate_data()`
 
   
 # rbmi 1.1.1 & 1.1.3
-  
-* Various minor tweaks to address CRAN checks
-* Otherwise no change in functionality from 1.1.0
+ 
+* No change in functionality from 1.1.0
+* Various minor tweaks to address CRAN checks messages
   
 # rbmi 1.1.0
   
