+++ .github/actions/build-src/action.yml
@@ -13,7 +13,7 @@ runs:
                 pkgs = installed.packages()[c("Rcpp", "rstan", "rstantools"), c("Version", "Package")]
             )
         sink()
-    - shell: Rscript {0}
+      shell: Rscript {0}
     
     # The default R/stanmodels.R was updated in v2.2.0 (11Apr2022)
     # Normally for older versions it would change itself on the fly
