+++ .github/workflows/on_push.yaml
@@ -38,6 +38,8 @@ jobs:
       
       - name: Build src
         uses: ./.github/actions/build-src
+        with:
+          refresh: false
       
       - name: Document Code
         run: |
