### 2022-05-18 09b4b58
add test for new analysis functions
