+++ tests/testthat/test-analysis_result.R
@@ -1,0 +1,231 @@
+# Test functions relavent to analysis_result class
+
+test_that("basic constructions of `analysis_result` work as expected", {
+
+    expect_general <- function(x) {
+        expect_s3_class(x, c('analysis_result', 'list'))
+        expect_type(x, 'list')
+        expect_named(x)
+    }
+
+    # with optional elements: df, meta
+    x <- analysis_result(name = 'trt',
+                         est = 1,
+                         se = 2,
+                         df = 3,
+                         meta = list(visit = 1))
+    expect_general(x)
+    expect_length(x, 5)
+    expect_equal(names(x), c('name', 'est', 'se', 'df', 'meta'))
+    expect_true(assertthat::has_attr(x, 'meta'))
+    expect_equal(x$name, 'trt')
+    expect_equal(x$df, 3)
+    expect_equal(x$meta, list(visit = 1))
+
+    # without optional elements
+    x <- analysis_result(name = 'trt',
+                         est = 1,
+                         se = 2,
+                         meta = list(visit = 1))
+    expect_general(x)
+    expect_length(x, 4)
+    expect_equal(names(x), c('name', 'est', 'se', 'meta'))
+    expect_true(has_attr(x, 'meta'))
+    expect_equal(x$name, 'trt')
+    expect_equal(x$meta, list(visit = 1))
+
+    # without optional elements
+    x <- analysis_result(name = 'trt',
+                         est = 1)
+    expect_general(x)
+    expect_length(x, 2)
+    expect_equal(names(x), c('name', 'est'))
+    expect_false(has_attr(x, 'meta'))
+    expect_equal(x$name, 'trt')
+    expect_equal(x$est, 1)
+
+    # special input: se = NA
+    x <- analysis_result(name = 'trt',
+                         est = 1,
+                         se = NA)
+    expect_general(x)
+
+    expect_length(x, 3)
+    expect_equal(names(x), c('name', 'est', 'se'))
+    expect_true(is.na(x$se))
+
+    # special input: df = NA
+    x <- analysis_result(name = 'trt',
+                         est = 1,
+                         se = 2,
+                         df = NA)
+    expect_general(x)
+    expect_length(x, 4)
+    expect_equal(names(x), c('name', 'est', 'se', 'df'))
+    expect_true(is.na(x$df))
+})
+
+
+test_that("incorrect constructions of analysis_result fail", {
+
+    # test parameter type
+    expect_error(
+        analysis_result(name = 1,
+                        est = 1,
+                        se = 2,
+                        df = 3,
+                        meta = list(visit = 1)),
+        "name of analysis_result is not character"
+    )
+
+    expect_error(
+        analysis_result(name = 'a',
+                        est = 'b',
+                        se = 2,
+                        df = 3,
+                        meta = list(visit = 1)),
+        "est of analysis_result is not numeric"
+    )
+
+    expect_error(
+        analysis_result(name = 'a',
+                        est = 1,
+                        se = list(),
+                        df = 3,
+                        meta = list(visit = 1)),
+        "se of analysis_result is not numeric or NA or NULL"
+    )
+
+    expect_error(
+        analysis_result(name = 'a',
+                        est = 1,
+                        se = 2,
+                        df = data.frame(),
+                        meta = list(visit = 1)),
+        "df of analysis_result is not numeric or NA or NULL"
+    )
+
+    expect_error(
+        analysis_result(name = 'a',
+                        est = 1,
+                        se = 2,
+                        df = 3,
[…]
