### 2022-05-19 a687a2c
modify test for original analysis functions
