+++ tests/testthat/test-analyse.R
@@ -10,8 +10,8 @@ test_that("basic constructions of `analysis` work as expected",{
 
     x <- as_analysis(
         results = list(
-            list(p1 = list("est" = 1)),
-            list(p1 = list("est" = 2))
+            list(analysis_result(name = 'p1', est = 1)), # A nested structure is necessary here. The top level is a full result list. The 2nd level is each imputation. The 3rd level is each individual analysis result inside each imputation.
+            list(analysis_result(name = 'p1', est = 2))
         ),
         method = method_condmean(n_samples = 1)
     )
@@ -20,8 +20,8 @@ test_that("basic constructions of `analysis` work as expected",{
 
     x <- as_analysis(
         results = list(
-            list(p1 = list("est" = 1)),
-            list(p1 = list("est" = 2))
+            list(analysis_result(name = 'p1', est = 1)),
+            list(analysis_result(name = 'p1', est = 2))
         ),
         method = method_condmean(type = "jackknife")
     )
@@ -30,8 +30,8 @@ test_that("basic constructions of `analysis` work as expected",{
 
     x <- as_analysis(
         results = list(
-            list(p1 = list("est" = 1, "df" = 4, "se" = 1)),
-            list(p1 = list("est" = 2, "df" = 3, "se" = 3))
+            list(analysis_result(name = 'p1', est = 1, df = 4, se = 1)),
+            list(analysis_result(name = 'p1', est = 2, df = 3, se = 3))
         ),
         method = method_bayes(n_samples = 2)
     )
@@ -40,8 +40,8 @@ test_that("basic constructions of `analysis` work as expected",{
 
     x <- as_analysis(
         results = list(
-            list(p1 = list("est" = 1, "df" = 4, "se" = 1)),
-            list(p1 = list("est" = 2, "df" = 3, "se" = 3))
+            list(analysis_result(name = 'p1', est = 1, df = 4, se = 1)),
+            list(analysis_result(name = 'p1', est = 2, df = 3, se = 3))
         ),
         method = method_approxbayes(n_samples = 2)
     )
@@ -50,8 +50,8 @@ test_that("basic constructions of `analysis` work as expected",{
 
     x <- as_analysis(
         results = list(
-            list(p1 = list("est" = 1, "df" = 4, "se" = NA)),
-            list(p1 = list("est" = 2, "df" = 3, "se" = NA))
+            list(analysis_result(name = 'p1',  est = 1,  df = 4, se = NA)),
+            list(analysis_result(name = 'p1',  est = 2,  df = 3, se = NA))
         ),
         method = method_bayes(n_samples = 2)
     )
@@ -60,13 +60,12 @@ test_that("basic constructions of `analysis` work as expected",{
 
     x <- as_analysis(
         results = list(
-            list(p1 = list("est" = 1, "df" = 4, "se" = NA)),
-            list(p1 = list("est" = 2, "df" = 3, "se" = NA))
+            list(analysis_result(name = 'p1', est = 1, df = 4, se = NA)),
+            list(analysis_result(name = 'p1', est = 2,  df = 3, se = NA))
         ),
         method = method_approxbayes(n_samples = 2)
     )
     expect_true(validate(x))
-
 })
 
 
@@ -76,8 +75,7 @@ test_that("incorrect constructions of as_analysis fail", {
     expect_error(
         as_analysis(
             results = list(
-                list(p1 = list("est" = 1)),
-                list(p1 = list("est" = 2))
+                list(analysis_result(name = 'p1', est = 1), analysis_result(name = 'p1', est = 2))
             ),
             method = method_condmean(n_samples = 2)
         ),
@@ -87,8 +85,8 @@ test_that("incorrect constructions of as_analysis fail", {
     expect_error(
         as_analysis(
             results = list(
[…]
