+++ NAMESPACE
@@ -39,7 +39,6 @@ export(add_class)
 export(analyse)
 export(analysis_result)
 export(ancova)
-export(as_analysis_result)
 export(as_class)
 export(as_vcov)
 export(delta_template)
