### 2022-05-20 bbdc170
update check in as_analysis
