### 2022-05-20 1a0d47c
update test for analysis functions
