### 2022-05-20 f0e88b1
fix wrong counting of analysis names
