### 2022-05-21 2f7956d
add test for util functions compose_n and back_apply_at
