### 2022-05-21 56e9b65
add extract analysis result function
