### 2022-05-22 302f351
add assertation to extract_analysis_result
