+++ R/analyse.R
@@ -737,6 +737,11 @@ analysis_info <- function(example, name_of_group = 'name', name_of_meta = 'meta'
 #' }
 extract_analysis_result <- function(results, ...){
     dots <- list(...)
+    assert_that(all(!is.null(names(dots)),
+                    length(names(dots)) > 0,
+                    !any(grepl("^$", names(dots)))),
+                msg = "Invalide parameters. Only key-word parameters are valide.")
+
     meta <- list()
     has_meta <- FALSE
     if ('meta' %in% names(dots)) {
@@ -746,9 +751,9 @@ extract_analysis_result <- function(results, ...){
     }
 
     names_match_values <- function(obj, named_values=dots) {
-        mapply(function(label, value) obj[[label]] == value,
+        mapply(function(label, value) isTRUE(obj[[label]] == value),
             names(named_values), named_values, SIMPLIFY = TRUE, USE.NAMES = FALSE)
-    }
+    } # When SIMPLIFY = TRUE, coercion can happen on logical(0) which generates WARNINGS. isTRUE is used to avoid coercion and make the code more robust
 
     extract_match <- function(obj, named_values=dots, constrain = identity) {
         Filter(function(item) all(
