### 2022-05-23 ebc556f
update all tests functions in test-ancova
