### 2022-05-24 b074bf8
make extract_analysis_result more robust
