### 2022-05-24 2496b6a
add test for extract_analysis_result
