### 2022-05-25 #4251 @fboehm
the dois in the pdf all resolve to the correct targets.
