### 2022-06-02 8a8d9f3
update doc for extract_analysis_result
