+++ man/extract_analysis_result.Rd
@@ -1,0 +1,39 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/analyse.R
+\name{extract_analysis_result}
+\alias{extract_analysis_result}
+\title{Extract analysis_result from a list of analysis_results by matching names and values}
+\usage{
+extract_analysis_result(results, ...)
+}
+\arguments{
+\item{results}{A list of analysis_result}
+
+\item{...}{Keywords parameters with the name and value matching the element in analysis result to be extracted}
+}
+\value{
+A list of matched analysis results
+}
+\description{
+The function returns a list of all analysis_results in the input list that match the values with names specified via keywords parameters of the function.
+If no value matches the specified name in any analysis_result containing in the given list or
+the specified name does not existed in any analysis_result, the function returns an empty list \code{list()}.
+}
+\examples{
+\dontrun{
+results <- list(
+    analysis_result(
+        name = 'trt',
+        est = 1,
+        se = 2,
+        df = 3,
+        meta = list(visit = 'vis1')
+    )
+)
+
+extract_analysis_result(results, name = 'trt')
+extract_analysis_result(results, est = 1)
+extract_analysis_result(results, name = 'trt', meta = list(visit = 'vis1'))
+extract_analysis_result(results, name = 'trt2')
+}
+}
