+++ man/transpose_results.Rd
@@ -19,37 +19,49 @@ the same estimates together into vectors.
 \details{
 Essentially this function takes an object of the format:\preformatted{x <- list(
     list(
-        "trt1" = list(
+     analysis_result(
+            name = 'trt',
             est = 1,
-            se  = 2
+            se  = 2,
+            meta = list(visit = 1)
         ),
-        "trt2" = list(
+     analysis_result(
+            name = 'trt',
             est = 3,
-            se  = 4
+            se  = 4,
+            meta = list(visit = 2)
         )
     ),
     list(
-        "trt1" = list(
+     analysis_result(
+            name = 'trt',
             est = 5,
-            se  = 6
+            se  = 6,
+            meta = list(visit = 1)
         ),
-        "trt2" = list(
-            est = 7,
-            se  = 8
+     analysis_result(
+           name = 'trt',
+           est = 7,
+           se  = 8,
+           meta = list(visit = 2)
         )
     )
 )
 }
 
 and produces:\preformatted{list(
-    trt1 = list(
+    trt.1 = data.frame(
+     list(
         est = c(1,5),
         se = c(2,6)
+        ),
     ),
-    trt2 = list(
+    trt.2 = data.frame(
+     list(
         est = c(3,7),
         se = c(4,8)
-    )
+        )
+     )
 )
 }
 }
