### 2022-06-02 cc9e859
update test for new pool function
