### 2022-06-13 #4251 @nociale
@fboehm Please apologize for the large delay (I was on annual leave for a while). I provide below the information about version number of the release and the doi of the archive:

- The archive: https://zenodo.org/record/6632154 (DOI: 10.5281/zenodo.6632154)
- The github release: https://github.com/insightsengineering/rbmi/releases/tag/v1.1.4 (Version: 1.1.4)

I hope we addressed the tasks correctly, please let me know whether the information provided are enough or we need to provide something more.

Thanks!
Kind regards,
Alessandro
