+++ R/pool.R
@@ -72,16 +72,18 @@ pool <- function(
 
     pool_type <- class(results$results)[[1]]
 
-    results_transpose <- transpose_results(results$results)
+    prepool <- make_poolable(results$results)
 
-    pars <- lapply(
-        results_transpose,
+    par_values <- lapply(
+        prepool$results,
         function(x, ...) pool_internal(as_class(x, pool_type), ...),
         conf.level = conf.level,
         alternative = alternative,
         type = type,
         D = results$method$D
     )
+
+    pars <- mapply(function(x, y) append(x, y), prepool$meta, par_values, SIMPLIFY = FALSE)
 
     if (pool_type == "bootstrap") {
         method <- sprintf("%s (%s)", pool_type, type)
@@ -94,7 +96,8 @@ pool <- function(
         conf.level = conf.level,
         alternative = alternative,
         N = length(results$results),
-        method = method
+        method = method,
+        metakeys = prepool$metakeys
     )
     class(ret) <- "pool"
     return(ret)
@@ -599,10 +602,10 @@ parametric_ci <- function(point, se, alpha, alternative, qfun, pfun, ...) {
 
 
 
-#' Transpose results object
+#' Convert analysis results to a poolable object
 #'
-#' Transposes a Results object (as created by [analyse()]) in order to group
-#' the same estimates together into vectors.
+#' Covert Results object (as created by [analyse()]) in order to group
+#' the same estimates together into vectors. The return object is in poolable class containing the results, meta information and the key names for the mata information
 #'
 #' @param results A list of results.
 #' @param non_group_keys a character vector of variables that are not used to group results usually variables of the numeric analysis results
@@ -610,7 +613,7 @@ parametric_ci <- function(point, se, alpha, alternative, qfun, pfun, ...) {
 #'
 #' @details
 #'
-#' Essentially this function takes an object of the format:
+#' The format of analysis results are converted from:
 #'
 #' ```
 #' x <- list(
@@ -645,46 +648,66 @@ parametric_ci <- function(point, se, alpha, alternative, qfun, pfun, ...) {
 #' )
 #' ```
 #'
-#' and produces:
+#' to the following format and stored in the `$$results` element of the poolable object. The element `$meta` contains meta information.
+#' The element `$metakey` contains the column names of the meta information as a character vector
 #'
 #' ```
 #' list(
 #'     trt.1 = data.frame(
 #'      list(
+#'         name = 'trt',
 #'         est = c(1,5),
-#'         se = c(2,6)
+#'         se = c(2,6),
+#'         visit = c(1,1)
 #'         ),
 #'     ),
 #'     trt.2 = data.frame(
 #'      list(
+#'         name = 'trt',
 #'         est = c(3,7),
-#'         se = c(4,8)
+#'         se = c(4,8),
+#'         visit = c(2,2)
 #'         )
 #'      )
 #' )
 #' ```
-transpose_results <- function(results, non_group_keys=c("est", "se", "df")) {
+make_poolable <- function(results, non_group_keys=c("est", "se", "df")) {
+    assert_map <- function(f, assert_f, logic = all) {
+        function(lst, ...) {
+            out <- lapply(lst, f, ...)
+            conds <- lapply(out, assert_f)
+            assert_that(logic(unlist(conds)))
+            out
+        }
+    }
+
+    extract_aggregate <- function(lst, keys) unique(lst[keys])
+    extract_aggragates <- assert_map(extract_aggregate, function(x) nrow(x) == 1)
+
[…]
