### 2022-06-13 a32ec18
add utility functions
