+++ man/vec2form.Rd
@@ -1,0 +1,19 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/utilities.R
+\name{vec2form}
+\alias{vec2form}
+\title{Convert vector to formula}
+\usage{
+vec2form(chr, bothside = FALSE)
+}
+\arguments{
+\item{chr}{character vector to be converted}
+
+\item{bothside}{A logical variable indicating whether to generate fomula with both right and left side. Default: FALSE - only right side formula will be generated}
+}
+\value{
+an object of formula class representing formula ~ chr[\link{1}] + chr[\link{2}] + ...
+}
+\description{
+Convert character vector c('a1', 'a2') to formula ~ a1 + a2
+}
