+++ man/make_poolable.Rd
@@ -1,0 +1,72 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/pool.R
+\name{make_poolable}
+\alias{make_poolable}
+\title{Convert analysis results to a poolable object}
+\usage{
+make_poolable(results, non_group_keys = c("est", "se", "df"))
+}
+\arguments{
+\item{results}{A list of results.}
+
+\item{non_group_keys}{a character vector of variables that are not used to group results usually variables of the numeric analysis results
+(Default: \verb{"est", "se", "df}).}
+}
+\description{
+Covert Results object (as created by \code{\link[=analyse]{analyse()}}) in order to group
+the same estimates together into vectors. The return object is in poolable class containing the results, meta information and the key names for the mata information
+}
+\details{
+The format of analysis results are converted from:\preformatted{x <- list(
+    list(
+     analysis_result(
+            name = 'trt',
+            est = 1,
+            se  = 2,
+            meta = list(visit = 1)
+        ),
+     analysis_result(
+            name = 'trt',
+            est = 3,
+            se  = 4,
+            meta = list(visit = 2)
+        )
+    ),
+    list(
+     analysis_result(
+            name = 'trt',
+            est = 5,
+            se  = 6,
+            meta = list(visit = 1)
+        ),
+     analysis_result(
+           name = 'trt',
+           est = 7,
+           se  = 8,
+           meta = list(visit = 2)
+        )
+    )
+)
+}
+
+to the following format and stored in the \verb{$results} element of the poolable object. The element \verb{$meta} contains meta information.
+The element \verb{$metakeys} contains the names of the meta information as a character vector\preformatted{list(
+    trt.1 = data.frame(
+     list(
+        name = 'trt',
+        est = c(1,5),
+        se = c(2,6),
+        visit = c(1,1)
+        ),
+    ),
+    trt.2 = data.frame(
+     list(
+        name = 'trt',
+        est = c(3,7),
+        se = c(4,8),
+        visit = c(2,2)
+        )
+     )
+)
+}
+}
