+++ man/transpose_results.Rd
@@ -1,67 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/pool.R
-\name{transpose_results}
-\alias{transpose_results}
-\title{Transpose results object}
-\usage{
-transpose_results(results, non_group_keys = c("est", "se", "df"))
-}
-\arguments{
-\item{results}{A list of results.}
-
-\item{non_group_keys}{a character vector of variables that are not used to group results usually variables of the numeric analysis results
-(Default: \verb{"est", "se", "df}).}
-}
-\description{
-Transposes a Results object (as created by \code{\link[=analyse]{analyse()}}) in order to group
-the same estimates together into vectors.
-}
-\details{
-Essentially this function takes an object of the format:\preformatted{x <- list(
-    list(
-     analysis_result(
-            name = 'trt',
-            est = 1,
-            se  = 2,
-            meta = list(visit = 1)
-        ),
-     analysis_result(
-            name = 'trt',
-            est = 3,
-            se  = 4,
-            meta = list(visit = 2)
-        )
-    ),
-    list(
-     analysis_result(
-            name = 'trt',
-            est = 5,
-            se  = 6,
-            meta = list(visit = 1)
-        ),
-     analysis_result(
-           name = 'trt',
-           est = 7,
-           se  = 8,
-           meta = list(visit = 2)
-        )
-    )
-)
-}
-
-and produces:\preformatted{list(
-    trt.1 = data.frame(
-     list(
-        est = c(1,5),
-        se = c(2,6)
-        ),
-    ),
-    trt.2 = data.frame(
-     list(
-        est = c(3,7),
-        se = c(4,8)
-        )
-     )
-)
-}
-}
