+++ NAMESPACE
@@ -45,6 +45,7 @@ export(delta_template)
 export(draws)
 export(expand)
 export(expand_locf)
+export(extract_analysis_result)
 export(extract_imputed_dfs)
 export(fill_locf)
 export(getStrategies)
