### 2022-06-14 #4251 @arfon
@nociale – I made a couple of minor tweaks to the paper here, could you merge them please? https://github.com/insightsengineering/rbmi/pull/367
