### 2022-06-15 #4251 @nociale
That is really a great news! :)

@arfon, @fboehm Thanks a lot for helping with the review and publication process!
@DanielRivasMD @JoranTiU Again, thanks for the review and for the positive feedback!

@arfon, @fboehm May I ask you about one little detail: The web title as currently appears in a Google search contains a typo. It is "rbmi: AR package for standard and reference-based multiple ...", while it should be "rbmi: A R package for standard and reference-based multiple ...", as correctly appears in the web page. Would it be possible to write this correctly? Thanks!
