### 2022-06-21 #4251 @nociale
@fboehm - The error is in the title I see on Google when I try to search for the paper (e.g. [here](https://www.google.com/search?q=rbmi%3A+A+R+package+for+standard+and+reference-based+multiple&rlz=1C1GCEA_enCH943CH943&ei=6xCwYvi3G4KTkwWR4L7wAw&ved=0ahUKEwi42bLxsLv4AhWCyaQKHRGwDz4Q4dUDCA4&uact=5&oq=rbmi%3A+A+R+package+for+standard+and+reference-based+multiple&gs_lcp=Cgdnd3Mtd2l6EANKBAhBGAFKBAhGGABQkFFYkmdgg2loAnAAeACAAdoBiAGjA5IBBTEuMS4xmAEAoAEBoAECsAEAwAEB&sclient=gws-wiz)). Not a big problem of course!
