### 2022-10-11 #4224 @JoranTiU
Hi all,

I can be the second reviewer on this 😊

Vriendelijke groet/Kind Regards,
Joran

Joran Jongerling | Assistant-Professor | Department of Methodology and Statistics | Tilburg School of Social and Behavioral Sciences | Simon Building, room S706

Van: Daniel Rivas ***@***.***>
Verzonden: donderdag 17 maart 2022 1:49
Aan: openjournals/joss-reviews ***@***.***>
CC: Joran Jongerling ***@***.***>; Mention ***@***.***>
Onderwerp: Re: [openjournals/joss-reviews] [PRE REVIEW]: rbmi: A R package for standard and reference-based multiple imputation methods (Issue #4224)

Sounds great! Let's do it! I read the article, and will install and test the software, is this the review thread where I should comment? Or where can I find it?
