### 2022-10-17 #4224 @JoranTiU
I hope so too 😊. And you’re very welcome 😊

Vriendelijke groet/Kind Regards,
Joran

Joran Jongerling | Assistant-Professor/Co-Director of TESC | Department of Methodology and Statistics | Tilburg School of Social and Behavioral Sciences | Reitse Poort, Room RP 15 | https://experiencesampling.nl/person/joran-jongerling/

***@***.***

From: Frederick Boehm ***@***.***>
Sent: zondag 16 oktober 2022 4:38
Cc: Joran Jongerling ***@***.***>; Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: rbmi: A R package for standard and reference-based multiple imputation methods (Issue #4224)

@JoranTiU<https://github.com/JoranTiU> - Thank you so much for the offer! We've actually completed the review for this submission. I hope that you and I can work together on a future submission. Thanks again!
